Add files via upload
The config file is the file he said to make for constants and the like
my page.py is assuming that when we create a page, we will specify the size of the objects that it will contain. Will need to be changed if we deem that not to be the case.